refactor : util 함수를 역할에 따라 두 개의 폴더(input,game)로 분리한다.
