+++ src/utils/game.js
@@ -1,4 +1,4 @@
-import { BASEBALL_RULE } from './constants.js';
+import { BASEBALL_RULE } from '../constants.js';
 
 export const generateRandomNumber = () => {
   const numberSet = new Set();
@@ -11,24 +11,6 @@ export const generateRandomNumber = () => {
     );
   }
   return [...numberSet].join('');
-};
-
-const hasZero = (target) => {
-  return target.indexOf('0') !== -1;
-};
-
-const isValidLength = (target) => {
-  return new Set(target.split('')).size === BASEBALL_RULE.DIGITS;
-};
-
-export const isNotValidInput = (input) => {
-  return (
-    !input || !isValidLength(input) || hasZero(input) || Number.isNaN(+input)
-  );
-};
-
-export const parseInput = (input) => {
-  return String(parseInt(input.trim(), 10));
 };
 
 const convertToHashMap = (input) => {
